efi/boot: Add "unified Xen" support for SecureBoot
During UEFI boot, Xen can look to see if it has built-in PE sections for its
configuration file, the Linux kernel, the initrd, and the XSM.  By
bundling all of these into a single file with the hypervisor, it is
easier to integrate with UEFI SecureBoot, which supports validating a single
signed EFI executable.

The PE executable parser and unifed kernel idea is copied from systemd-boot
and has been used by the safeboot project.

The scripts/unify-xen shows how to use objcopy to bundle all of the
components together; it requires better command line parsing to be
robust.

Signed-off-by: Trammell Hudson <hudson@trmm.net>